requirements.txt
fastapi
asyncpg
websockets
python-dotenv
aiohttp
bcrypt
snowflake-id
pyotp
qrcode[pil]